run_qemu: add an option to boot the rootfs as-is
Add an option to run_qemu to let it boot the rootfs image directly
instead of providing a kernel image on the qemu command line (i.e.
qemu's '-kernel' option). This requires a correctly setup EFI bootloader
in the image, configured with the desired kernel to boot.

The run_qemu option added is --direct-kernel, which defaults to 'on'. To
negate this, use --no-direct-kernel. This is needed because defining a
negative opt, such as --no-kernel, doesn't work as expected with
argbash.

Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>