run_qemu: Correct kernel image names
run_qemu has multiple places that depend on the specific kernel image
being named vmlinuz-"kver". This is fine if your host distro has an
installkernel script that does this, or the user has manually installed
such a script. For all other cases however, the kernel image will be
named something else.

While the existing code did the right thing as a fallback for qemu, the
code which creates a real boot partition will fail.

Signed-off-by: Ben Widawsky <ben.widawsky@intel.com>