run_qemu: update --cxl to qemu upstream
This updates run_qemu's qemu command line generation to the CXL patches
that have been merged upstream. It includes the cxl-fmw reworks that
make the memory windows a machine parameter. This was merged as of:

  30796f556790 ("Merge tag 'for_upstream' of git://git.kernel.org/pub/scm/virt/kvm/mst/qemu into staging")

It is recommended that users update their qemu trees from developer
branches to tip of the upstream tree at this point.

Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>